Revert "Remove views"
