+++ src/api/views/__init__.py
@@ -1,4 +1,4 @@
 """
 The views package
 """
-from api.views import oauthView
+from api.views import userView, projectView, oauthView
